Move non-CLI tests into `test/splitgraph/cloud`.
