+++ test/splitgraph/cloud/project/snapshots/test_generation/test_generate_project_no_dbt/generate_project/splitgraph.yml
